+++ tests/eth2/fixtures/test_bls.py
@@ -1,4 +1,4 @@
-from eth2.beacon.tools.fixtures.test_gen import (
+from eth2.beacon.tools.fixtures.test_generation import (
     generate_pytests_from_eth2_fixture,
     pytest_from_eth2_fixture,
 )
