+++ tests/eth2/fixtures/test_epoch_processing.py
@@ -1,5 +1,5 @@
 from eth2.beacon.tools.fixtures.config_types import Minimal
-from eth2.beacon.tools.fixtures.test_gen import (
+from eth2.beacon.tools.fixtures.test_generation import (
     generate_pytests_from_eth2_fixture,
     pytest_from_eth2_fixture,
 )
